Implementing Camera for Sale Options Product Photos
The user can now take photos in addition to selecting photos from the photo library.  ProductImageCollectionViewController uses notifications to communicate with the BroadcastViewController when the camera is selected and dismissed.  SaleOptionsViewController now has BroadcastViewController as its delegate so that we can dismiss the BroadcastViewController under the SaleOptionsViewController if the user presses the cancel "X" icon in the SaleOptionsViewcontroller.