Checking if FBUser is null
If user is null, then this is the first time logging in.  We should get the user's name info from the AWS provider's info (in this case twitter).  Otherwise, user info should be populated from firebase.  We were checking for "nil" instead of "null" but firebase return "null" for snapshot.value if there is no data.