Saving Broadcast images with S3Handler and then saving broadcast
Broadcast images are now handled with the S3Handler.  S3Handler will be in charge up uploading images to AWS S3.  After images are uploaded, the SaleOptionsViewController notifies the BroadcastViewController to begin publishing.  Once publishing begins, the broadcast object will be saved to firebase.  We have a delay of 2 seconds before saving the broadcast object to firebase.  This gives the S3Handler time to upload the images to AWS and allow for the Fastly CDN to properly point to the cached images.