Removing filter for publisher's broadcasts
We were going to filter out the publisher's broadcasts so that they do not see their own broadcasts.  This would be tricky with Firebase-UI, we do not want to mess with the array that is synced real time.  Instead, we will come up with a way to subscribe only to broadcasts that are not the users.  We will create another url in firbase that contains all broadcasts except for the publisher's broadcasts.  The DiscoverCollectionViewController will then point to this url instead.  As a result, the publisher's broadcasts will not appear in the collection view.