Linking up subscriberCount and quantity.  Adding Buy button
Subscriber Count and the Quantity values on the ChatViewController are now synced with firebase.  The Broadcast object uses a delegate to update the quantity and subscriber count when the broadcast receives an update.  Added a buy button that we will use later to decrement counter when subscriber buys an item.