Refactoring ChatViewController and DetailViewController
ChatViewController and DetailViewController were being setup in the SubscribeOverlayViewController.  Now there is OverlayPageViewController that handles this.  The countdown timer and the quantity label have been moved to the SubscribeOverlayViewController from the ChatViewController so that it remains on the screen when the user swipes between the chatViewController and the DetailViewController.  The countdown timer is now on the NSRunLoopCommonModes so that it continues to countdown even when the user is interacting with the screen.