Adding Messaging to BroadcastOverlayViewController
BroadcastOverlayViewController now registers and listens for messages on the broadcast.  Moved the initialization of the BroadcastOverlayViewController and SubscriberOverlayViewController to the respective viewDidLoad methods.  They were being initialized in viewWillLayoutSubviews, which happens more than once.  Firebase childAdded observer shows all initial messages at first.  Added a workaround for this by populating the messages for a broadcast with an initial message and then ignoring the first message after registering for the message notifications in the app.  This way, the first message primes the observers, but the first message is not displayed to the screen.