DetailViewController adds subscribers to broadcast.
The DetailViewController will now display the subscribers to a broadcast.  This is implemented using DetailCollectionViewDataSource, which is a subclass of FirebaseCollectionViewDataSource.  The SubscriberViewController now saves a subscriber to the firebase subscribers url when the subscriber session connects.  This needs to be updated to save when the subscriber stream connects, but we are using the session connect for testing since we don't have another device to create the stream connection.  The broadcast also increments the subscriber count in firebase when the subscriber is added.